Fixes hardcoded LB IP address pool
TL;DR
-----

Repairs a defect where the specified balancer CIDR was ignored

Details
-------

Uses the expected parameter to set the CIDR for load balancer IPs.
In the current code the value is hardcodeed for the old CIDR
used for the ferret cluster so all clusters were getting that IP
range. Once this change is applied the parameter is used.