Implements wasm as an optional runtime
TL;DR
-----

Provides a wasm runtime when `cluster.runtimes.wasm.enabled`
is set to true

Details
-------

Uses the K0s wasm enabler to install a wasm runtime and
create a runtime class for it. The installation is conditioned on
a variable in `secrets/params.yaml` being set to true.

To include wasm in your cluster, add

```
cluster:
  runtimes:
    wasm:
      enabled: true
```

to your parameters file.