Restores DHCP for workload interface
TL;DR
-----

Corrects issue where MetalLB couldn't advertise addresses

Details
-------

Fixes issue where addresses for `LoadBalancer` services weren't
being advertised correctly. The root cause was that when I should
have only removed the routing table for the `10.26.0.0/16` subnet,
I instead removed all of the configuration for the second network
interface. This meant the workers didn't have an address on the
workload network so there was not way to route traffic.