Updates vCenter key thumbprint
TL;DR
-----

Assures the right thumbprint for the vCenter certificate

Details
-------

The vSphere CSI stopped working on my clusters after I rotated the
certificate for vCenter. This change updates the thumbprint in the
template the creates the vSphere CSI configuration secret so that it
can properly validate the connection when the server presents the new
certificate.