Adapts for Nutanix
TL;DR
-----

Creates nodes on Nutanix instead of vSphere

Details
-------

Converts the cluster infrastructure to run on Nutanix instead of
vSphere. Sorry, Broadcom.

This is a response to the end of VMUG licensing. I've converted my
homelabe from vSphere to Nutanix CE becasue the price (free as in
beer) is just right.

All the Terraform is updated to work with vSphere and instead of
installing the vSphere CSI it installs the Nutanix one. The CSI
will likely need some more configuration once I get a more complete
Nutanix implementation, this was a toy to make sure I could convert
at all.